@@ -12,12 +12,11 @@
         }
     ],
     "require": {
-        "php": "^7.1",
+        "php": "^7.3",
         "symfony/expression-language": "^4"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7 || ^8",
-        "friendsofphp/php-cs-fixer": "^2"
+        "phpunit/phpunit": "^8"
     },
     "autoload": {
         "psr-4": {
