@@ -2,7 +2,6 @@
 
 namespace uuf6429\ExpressionLanguage;
 
-use uuf6429\ExpressionLanguage\Node\ArrowFuncNode;
 use Symfony\Component\ExpressionLanguage\Node\ArgumentsNode;
 use Symfony\Component\ExpressionLanguage\Node\ArrayNode;
 use Symfony\Component\ExpressionLanguage\Node\BinaryNode;
@@ -13,10 +12,11 @@ use Symfony\Component\ExpressionLanguage\Node\GetAttrNode;
 use Symfony\Component\ExpressionLanguage\Node\NameNode;
 use Symfony\Component\ExpressionLanguage\Node\Node;
 use Symfony\Component\ExpressionLanguage\Node\UnaryNode;
-use Symfony\Component\ExpressionLanguage\Token;
-use Symfony\Component\ExpressionLanguage\SyntaxError;
 use Symfony\Component\ExpressionLanguage\Parser as SymfonyParser;
+use Symfony\Component\ExpressionLanguage\SyntaxError;
+use Symfony\Component\ExpressionLanguage\Token;
 use Symfony\Component\ExpressionLanguage\TokenStream as SymfonyTokenStream;
+use uuf6429\ExpressionLanguage\Node\ArrowFuncNode;
 
 class Parser extends SymfonyParser
 {
@@ -49,43 +49,43 @@ class Parser extends SymfonyParser
 
         $this->functions = $functions;
 
-        $this->unaryOperators = array(
-            'not' => array('precedence' => 50),
-            '!' => array('precedence' => 50),
-            '-' => array('precedence' => 500),
-            '+' => array('precedence' => 500),
-        );
-        $this->binaryOperators = array(
-            '->' => array('precedence' => 5, 'associativity' => self::OPERATOR_LEFT),
-            'or' => array('precedence' => 10, 'associativity' => self::OPERATOR_LEFT),
-            '||' => array('precedence' => 10, 'associativity' => self::OPERATOR_LEFT),
-            'and' => array('precedence' => 15, 'associativity' => self::OPERATOR_LEFT),
-            '&&' => array('precedence' => 15, 'associativity' => self::OPERATOR_LEFT),
-            '|' => array('precedence' => 16, 'associativity' => self::OPERATOR_LEFT),
-            '^' => array('precedence' => 17, 'associativity' => self::OPERATOR_LEFT),
-            '&' => array('precedence' => 18, 'associativity' => self::OPERATOR_LEFT),
-            '==' => array('precedence' => 20, 'associativity' => self::OPERATOR_LEFT),
-            '===' => array('precedence' => 20, 'associativity' => self::OPERATOR_LEFT),
-            '!=' => array('precedence' => 20, 'associativity' => self::OPERATOR_LEFT),
-            '!==' => array('precedence' => 20, 'associativity' => self::OPERATOR_LEFT),
-            '<' => array('precedence' => 20, 'associativity' => self::OPERATOR_LEFT),
-            '>' => array('precedence' => 20, 'associativity' => self::OPERATOR_LEFT),
-            '>=' => array('precedence' => 20, 'associativity' => self::OPERATOR_LEFT),
-            '<=' => array('precedence' => 20, 'associativity' => self::OPERATOR_LEFT),
-            'not in' => array('precedence' => 20, 'associativity' => self::OPERATOR_LEFT),
-            'in' => array('precedence' => 20, 'associativity' => self::OPERATOR_LEFT),
-            'matches' => array('precedence' => 20, 'associativity' => self::OPERATOR_LEFT),
-            '..' => array('precedence' => 25, 'associativity' => self::OPERATOR_LEFT),
-            '+' => array('precedence' => 30, 'associativity' => self::OPERATOR_LEFT),
-            '-' => array('precedence' => 30, 'associativity' => self::OPERATOR_LEFT),
-            '~' => array('precedence' => 40, 'associativity' => self::OPERATOR_LEFT),
-            '*' => array('precedence' => 60, 'associativity' => self::OPERATOR_LEFT),
-            '/' => array('precedence' => 60, 'associativity' => self::OPERATOR_LEFT),
-            '%' => array('precedence' => 60, 'associativity' => self::OPERATOR_LEFT),
-            '**' => array('precedence' => 200, 'associativity' => self::OPERATOR_RIGHT),
-        );
+        $this->unaryOperators = [
+            'not' => ['precedence' => 50],
+            '!' => ['precedence' => 50],
+            '-' => ['precedence' => 500],
+            '+' => ['precedence' => 500],
+        ];
+        $this->binaryOperators = [
+            '->' => ['precedence' => 5, 'associativity' => self::OPERATOR_LEFT],
+            'or' => ['precedence' => 10, 'associativity' => self::OPERATOR_LEFT],
+            '||' => ['precedence' => 10, 'associativity' => self::OPERATOR_LEFT],
+            'and' => ['precedence' => 15, 'associativity' => self::OPERATOR_LEFT],
+            '&&' => ['precedence' => 15, 'associativity' => self::OPERATOR_LEFT],
+            '|' => ['precedence' => 16, 'associativity' => self::OPERATOR_LEFT],
+            '^' => ['precedence' => 17, 'associativity' => self::OPERATOR_LEFT],
+            '&' => ['precedence' => 18, 'associativity' => self::OPERATOR_LEFT],
+            '==' => ['precedence' => 20, 'associativity' => self::OPERATOR_LEFT],
+            '===' => ['precedence' => 20, 'associativity' => self::OPERATOR_LEFT],
+            '!=' => ['precedence' => 20, 'associativity' => self::OPERATOR_LEFT],
+            '!==' => ['precedence' => 20, 'associativity' => self::OPERATOR_LEFT],
+            '<' => ['precedence' => 20, 'associativity' => self::OPERATOR_LEFT],
+            '>' => ['precedence' => 20, 'associativity' => self::OPERATOR_LEFT],
+            '>=' => ['precedence' => 20, 'associativity' => self::OPERATOR_LEFT],
+            '<=' => ['precedence' => 20, 'associativity' => self::OPERATOR_LEFT],
+            'not in' => ['precedence' => 20, 'associativity' => self::OPERATOR_LEFT],
+            'in' => ['precedence' => 20, 'associativity' => self::OPERATOR_LEFT],
+            'matches' => ['precedence' => 20, 'associativity' => self::OPERATOR_LEFT],
+            '..' => ['precedence' => 25, 'associativity' => self::OPERATOR_LEFT],
+            '+' => ['precedence' => 30, 'associativity' => self::OPERATOR_LEFT],
+            '-' => ['precedence' => 30, 'associativity' => self::OPERATOR_LEFT],
+            '~' => ['precedence' => 40, 'associativity' => self::OPERATOR_LEFT],
+            '*' => ['precedence' => 60, 'associativity' => self::OPERATOR_LEFT],
+            '/' => ['precedence' => 60, 'associativity' => self::OPERATOR_LEFT],
+            '%' => ['precedence' => 60, 'associativity' => self::OPERATOR_LEFT],
+            '**' => ['precedence' => 200, 'associativity' => self::OPERATOR_RIGHT],
+        ];
 
-        $this->replacementNodes = array();
+        $this->replacementNodes = [];
     }
 
     /**
@@ -108,7 +108,7 @@ class Parser extends SymfonyParser
      *
      * @throws SyntaxError
      */
-    public function parse(SymfonyTokenStream $stream, $names = array())
+    public function parse(SymfonyTokenStream $stream, $names = [])
     {
         $this->names = $names;
 
@@ -118,14 +118,7 @@ class Parser extends SymfonyParser
         $node = $this->parseExpression();
 
         if (!$stream->isEOF()) {
-            throw new SyntaxError(
-                sprintf(
-                    'Unexpected token "%s" of value "%s"',
-                    $stream->current->type,
-                    $stream->current->value
-                ),
-                $stream->current->cursor
-            );
+            throw new SyntaxError(sprintf('Unexpected token "%s" of value "%s"', $stream->current->type, $stream->current->value), $stream->current->cursor);
         }
 
         return $node;
@@ -133,8 +126,6 @@ class Parser extends SymfonyParser
 
     /**
      * Replaces all anonymous functions with placeholder tokens.
-     *
-     * @param TokenStream $stream
      *
      * @return TokenStream
      */
@@ -148,8 +139,8 @@ class Parser extends SymfonyParser
                 $this->replacementNodes[$replacementNodeIndex] = null;
 
                 // parse parameters
-                $parameterNames = array();
-                $parameterNodes = array();
+                $parameterNames = [];
+                $parameterNodes = [];
                 $expectParam = true;
                 $stream->prev();
                 $stream->expectPrev(Token::PUNCTUATION_TYPE, ')', 'Parameter list must end with parenthesis');
@@ -170,9 +161,9 @@ class Parser extends SymfonyParser
                 $stream->seek($operatorPos, SEEK_SET);
                 $stream->next();
                 $stream->expect(Token::PUNCTUATION_TYPE, '{', 'Anonymous function body must start with a curly bracket');
-                $bodyTokens = array();
+                $bodyTokens = [];
                 $openingBracketCount = 1;
-                while ($openingBracketCount != 0) {
+                while (0 != $openingBracketCount) {
                     if ($stream->current->test(Token::PUNCTUATION_TYPE, '{')) {
                         ++$openingBracketCount;
                     }
@@ -205,7 +196,7 @@ class Parser extends SymfonyParser
                 // update token stream
                 $this->replacementNodes[$replacementNodeIndex] = new ArrowFuncNode($parameterNodes, $bodyNode);
                 $replacement = new Token(static::TOKEN_REPLACEMENT_TYPE, $replacementNodeIndex, $operatorCursor);
-                $stream = $stream->splice($startPos, $endPos - $startPos, array($replacement));
+                $stream = $stream->splice($startPos, $endPos - $startPos, [$replacement]);
 
                 // keep parsing anonymous functions
                 $stream->seek($startPos, SEEK_SET);
@@ -423,14 +414,14 @@ class Parser extends SymfonyParser
     public function parsePostfixExpression($node): Node
     {
         $token = $this->stream->current;
-        while ($token->type === Token::PUNCTUATION_TYPE) {
+        while (Token::PUNCTUATION_TYPE === $token->type) {
             if ('.' === $token->value) {
                 $this->stream->next();
                 $token = $this->stream->current;
                 $this->stream->next();
 
                 if (
-                    $token->type !== Token::NAME_TYPE
+                    Token::NAME_TYPE !== $token->type
                     &&
                     // Operators like "not" and "matches" are valid method or property names,
 
@@ -443,7 +434,7 @@ class Parser extends SymfonyParser
                     // Other types, such as STRING_TYPE and NUMBER_TYPE, can't be parsed as property nor method names.
 
                     // As a result, if $token is NOT an operator OR $token->value is NOT a valid property or method name, an exception shall be thrown.
-                    ($token->type !== Token::OPERATOR_TYPE || !preg_match('/[a-zA-Z_\x7f-\xff][a-zA-Z0-9_\x7f-\xff]*/A', $token->value))
+                    (Token::OPERATOR_TYPE !== $token->type || !preg_match('/[a-zA-Z_\x7f-\xff][a-zA-Z0-9_\x7f-\xff]*/A', $token->value))
                 ) {
                     throw new SyntaxError('Expected name', $token->cursor);
                 }
@@ -482,7 +473,7 @@ class Parser extends SymfonyParser
      */
     public function parseArguments(): Node
     {
-        $args = array();
+        $args = [];
         $this->stream->expect(Token::PUNCTUATION_TYPE, '(', 'A list of arguments must begin with an opening parenthesis');
         while (!$this->stream->current->test(Token::PUNCTUATION_TYPE, ')')) {
             if (!empty($args)) {
@@ -496,9 +487,6 @@ class Parser extends SymfonyParser
         return new Node($args);
     }
 
-    /**
-     * @param array $functions
-     */
     public function setFunctions(array $functions): void
     {
         $this->functions = $functions;
