@@ -3,18 +3,18 @@
 namespace uuf6429\ExpressionLanguage\Tests;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\ExpressionLanguage\Node;
 use Symfony\Component\ExpressionLanguage\SyntaxError;
-use uuf6429\ExpressionLanguage\Parser;
 use uuf6429\ExpressionLanguage\Lexer;
 use uuf6429\ExpressionLanguage\Node\ArrowFuncNode;
-use Symfony\Component\ExpressionLanguage\Node;
+use uuf6429\ExpressionLanguage\Parser;
 
 class ParserTest extends TestCase
 {
     public function testParseWithInvalidName(): void
     {
         $lexer = new Lexer();
-        $parser = new Parser(array());
+        $parser = new Parser([]);
 
         $this->expectException(SyntaxError::class);
         $this->expectExceptionMessage('Variable "foo" is not valid around position 1.');
@@ -25,12 +25,12 @@ class ParserTest extends TestCase
     public function testParseWithZeroInNames()
     {
         $lexer = new Lexer();
-        $parser = new Parser(array());
+        $parser = new Parser([]);
 
         $this->expectException(SyntaxError::class);
         $this->expectExceptionMessage('Variable "foo" is not valid around position 1.');
 
-        $parser->parse($lexer->tokenize('foo'), array(0));
+        $parser->parse($lexer->tokenize('foo'), [0]);
     }
 
     /**
@@ -54,63 +54,63 @@ class ParserTest extends TestCase
         $arguments->addElement(new Node\ConstantNode(2));
         $arguments->addElement(new Node\ConstantNode(true));
 
-        return array(
-            array(
+        return [
+            [
                 new Node\NameNode('a'),
                 'a',
-                array('a'),
-            ),
-            array(
+                ['a'],
+            ],
+            [
                 new Node\ConstantNode('a'),
                 '"a"',
-            ),
-            array(
+            ],
+            [
                 new Node\ConstantNode(3),
                 '3',
-            ),
-            array(
+            ],
+            [
                 new Node\ConstantNode(false),
                 'false',
-            ),
-            array(
+            ],
+            [
                 new Node\ConstantNode(true),
                 'true',
-            ),
-            array(
+            ],
+            [
                 new Node\ConstantNode(null),
                 'null',
-            ),
-            array(
+            ],
+            [
                 new Node\UnaryNode('-', new Node\ConstantNode(3)),
                 '-3',
-            ),
-            array(
+            ],
+            [
                 new Node\BinaryNode('-', new Node\ConstantNode(3), new Node\ConstantNode(3)),
                 '3 - 3',
-            ),
-            array(
+            ],
+            [
                 new Node\BinaryNode('*',
                     new Node\BinaryNode('-', new Node\ConstantNode(3), new Node\ConstantNode(3)),
                     new Node\ConstantNode(2)
                 ),
                 '(3 - 3) * 2',
-            ),
-            array(
+            ],
+            [
                 new Node\GetAttrNode(new Node\NameNode('foo'), new Node\NameNode('bar'), new Node\ArgumentsNode(), Node\GetAttrNode::PROPERTY_CALL),
                 'foo.bar',
-                array('foo'),
-            ),
-            array(
+                ['foo'],
+            ],
+            [
                 new Node\GetAttrNode(new Node\NameNode('foo'), new Node\NameNode('bar'), new Node\ArgumentsNode(), Node\GetAttrNode::METHOD_CALL),
                 'foo.bar()',
-                array('foo'),
-            ),
-            array(
+                ['foo'],
+            ],
+            [
                 new Node\GetAttrNode(new Node\NameNode('foo'), new Node\NameNode('not'), new Node\ArgumentsNode(), Node\GetAttrNode::METHOD_CALL),
                 'foo.not()',
-                array('foo'),
-            ),
-            array(
+                ['foo'],
+            ],
+            [
                 new Node\GetAttrNode(
                     new Node\NameNode('foo'),
                     new Node\NameNode('bar'),
@@ -118,24 +118,24 @@ class ParserTest extends TestCase
                     Node\GetAttrNode::METHOD_CALL
                 ),
                 'foo.bar("arg1", 2, true)',
-                array('foo'),
-            ),
-            array(
+                ['foo'],
+            ],
+            [
                 new Node\GetAttrNode(new Node\NameNode('foo'), new Node\ConstantNode(3), new Node\ArgumentsNode(), Node\GetAttrNode::ARRAY_CALL),
                 'foo[3]',
-                array('foo'),
-            ),
-            array(
+                ['foo'],
+            ],
+            [
                 new Node\ConditionalNode(new Node\ConstantNode(true), new Node\ConstantNode(true), new Node\ConstantNode(false)),
                 'true ? true : false',
-            ),
-            array(
+            ],
+            [
                 new Node\BinaryNode('matches', new Node\ConstantNode('foo'), new Node\ConstantNode('/foo/')),
                 '"foo" matches "/foo/"',
-            ),
+            ],
 
             // chained calls
-            array(
+            [
                 $this->createGetAttrNode(
                     $this->createGetAttrNode(
                         $this->createGetAttrNode(
@@ -144,21 +144,21 @@ class ParserTest extends TestCase
                         'baz', Node\GetAttrNode::PROPERTY_CALL),
                     '3', Node\GetAttrNode::ARRAY_CALL),
                 'foo.bar().foo().baz[3]',
-                array('foo'),
-            ),
+                ['foo'],
+            ],
 
-            array(
+            [
                 new Node\NameNode('foo'),
                 'bar',
-                array('foo' => 'bar'),
-            ),
+                ['foo' => 'bar'],
+            ],
 
-            array(
+            [
                 new ArrowFuncNode(
-                    array(
+                    [
                         new Node\NameNode('foo'),
                         new Node\NameNode('bar'),
-                    ),
+                    ],
                     new Node\BinaryNode(
                         '*',
                         new Node\NameNode('foo'),
@@ -166,42 +166,42 @@ class ParserTest extends TestCase
                     )
                 ),
                 '(foo, bar) -> { foo * bar }',
-            ),
+            ],
 
-            array(
+            [
                 new ArrowFuncNode(
-                    array(
+                    [
                         new Node\NameNode('foo'),
                         new Node\NameNode('bars'),
-                    ),
+                    ],
                     new Node\BinaryNode(
                         '*',
                         new Node\NameNode('foo'),
                         new Node\FunctionNode(
                             'map',
                             new Node\Node(
-                                array(
+                                [
                                     new Node\NameNode('bars'),
                                     new ArrowFuncNode(
-                                        array(
+                                        [
                                             new Node\NameNode('bar'),
-                                        ),
+                                        ],
                                         new Node\BinaryNode(
                                             '*',
                                             new Node\NameNode('bar'),
                                             new Node\NameNode('baz')
                                         )
                                     ),
-                                )
+                                ]
                             )
                         )
                     )
                 ),
                 '(foo, bars) -> { foo * map(bars, (bar) -> { bar * baz }) }',
-                array('baz'),
-                array('map' => 'array_map'),
-            ),
-        );
+                ['baz'],
+                ['map' => 'array_map'],
+            ],
+        ];
     }
 
     private function createGetAttrNode($node, $item, $type): Node\GetAttrNode
@@ -214,10 +214,10 @@ class ParserTest extends TestCase
     /**
      * @dataProvider getInvalidPostfixData
      */
-    public function testParseWithInvalidPostfixData($expr, $names = array())
+    public function testParseWithInvalidPostfixData($expr, $names = [])
     {
         $lexer = new Lexer();
-        $parser = new Parser(array());
+        $parser = new Parser([]);
 
         $this->expectException(SyntaxError::class);
 
@@ -226,23 +226,23 @@ class ParserTest extends TestCase
 
     public function getInvalidPostfixData(): array
     {
-        return array(
-            array(
+        return [
+            [
                 'foo."#"',
-                array('foo'),
-            ),
-            array(
+                ['foo'],
+            ],
+            [
                 'foo."bar"',
-                array('foo'),
-            ),
-            array(
+                ['foo'],
+            ],
+            [
                 'foo.**',
-                array('foo'),
-            ),
-            array(
+                ['foo'],
+            ],
+            [
                 'foo.123',
-                array('foo'),
-            ),
-        );
+                ['foo'],
+            ],
+        ];
     }
 }
