@@ -3,15 +3,14 @@
 namespace uuf6429\ExpressionLanguage\Tests;
 
 use PHPUnit\Framework\TestCase;
-use uuf6429\ExpressionLanguage\Lexer;
 use Symfony\Component\ExpressionLanguage\Token;
+use uuf6429\ExpressionLanguage\Lexer;
 use uuf6429\ExpressionLanguage\TokenStream;
 
 class LexerTest extends TestCase
 {
     /**
      * @param Token[] $tokens
-     * @param string $expression
      *
      * @dataProvider getTokenizeData
      */
@@ -24,33 +23,33 @@ class LexerTest extends TestCase
 
     public function getTokenizeData(): array
     {
-        return array(
-            array(
-                array(new Token('name', 'a', 3)),
+        return [
+            [
+                [new Token('name', 'a', 3)],
                 '  a  ',
-            ),
-            array(
-                array(new Token('name', 'a', 1)),
+            ],
+            [
+                [new Token('name', 'a', 1)],
                 'a',
-            ),
-            array(
-                array(new Token('string', 'foo', 1)),
+            ],
+            [
+                [new Token('string', 'foo', 1)],
                 '"foo"',
-            ),
-            array(
-                array(new Token('number', '3', 1)),
+            ],
+            [
+                [new Token('number', '3', 1)],
                 '3',
-            ),
-            array(
-                array(new Token('operator', '+', 1)),
+            ],
+            [
+                [new Token('operator', '+', 1)],
                 '+',
-            ),
-            array(
-                array(new Token('punctuation', '.', 1)),
+            ],
+            [
+                [new Token('punctuation', '.', 1)],
                 '.',
-            ),
-            array(
-                array(
+            ],
+            [
+                [
                     new Token('punctuation', '(', 1),
                     new Token('number', '3', 2),
                     new Token('operator', '+', 4),
@@ -66,23 +65,23 @@ class LexerTest extends TestCase
                     new Token('punctuation', '[', 25),
                     new Token('number', '4', 26),
                     new Token('punctuation', ']', 27),
-                ),
+                ],
                 '(3 + 5) ~ foo("bar").baz[4]',
-            ),
-            array(
-                array(new Token('operator', '..', 1)),
+            ],
+            [
+                [new Token('operator', '..', 1)],
                 '..',
-            ),
-            array(
-                array(new Token('string', '#foo', 1)),
+            ],
+            [
+                [new Token('string', '#foo', 1)],
                 "'#foo'",
-            ),
-            array(
-                array(new Token('string', '#foo', 1)),
+            ],
+            [
+                [new Token('string', '#foo', 1)],
                 '"#foo"',
-            ),
-            array(
-                array(
+            ],
+            [
+                [
                     new Token('name', 'foo', 1),
                     new Token('punctuation', '(', 4),
                     new Token('punctuation', '(', 5),
@@ -95,9 +94,9 @@ class LexerTest extends TestCase
                     new Token('name', 'baz', 21),
                     new Token('punctuation', '}', 25),
                     new Token('punctuation', ')', 26),
-                ),
+                ],
                 'foo((bar, baz) -> { baz })',
-            ),
-        );
+            ],
+        ];
     }
 }
